Recipe Repository
moved the header/footer in allRecipes.js to _app.js. created getRecipe.js in the components folder and created RecipeCard.js in category folder. renamed the RecipeContainer to GetRecipe and replaced it from allRecipes.js to GetRecipe.js. function for button in the GetRecipe.js that will display the complete details in the RecipeCard.js page.